+++ core-commonui/unit-saveditems/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/saveditems/di/Utils.kt
@@ -1,0 +1,11 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.saveditems.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.saveditems.SavedItemsMviModel
+import org.koin.java.KoinJavaComponent
+
+actual fun getSavedItemsViewModel(): SavedItemsMviModel {
+    val res: SavedItemsMviModel by KoinJavaComponent.inject(
+        clazz = SavedItemsMviModel::class.java,
+    )
+    return res
+}
