+++ core-commonui/unit-saveditems/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/saveditems/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.saveditems.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.saveditems.SavedItemsMviModel
+
+expect fun getSavedItemsViewModel(): SavedItemsMviModel
