+++ settings.gradle.kts
@@ -54,3 +54,4 @@ include(":core-commonui:unit-remove")
 include(":core-commonui:unit-reportlist")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
+include(":core-commonui:unit-saveditems")
